Suppress a silly warning related to printing a time_t, the C standard is still undefined as to what a time_t is.  So just convert it to an int in this case.

+++ src/net/end_to_end_init.c
@@ -206,7 +206,7 @@ xdd_e2e_setup_src_socket(ptds_t *qp) {
 		req.tv_sec = 3;
 		fprintf(xgp->errout,
 			"Socket connection error, retrying in %d seconds: %d\n",
-			req.tv_sec, status);
+			(int)req.tv_sec, status);
 		nanosleep(&req, (struct timespec *)NULL);
 	    }
